feat(tekton): graduate NFS plugin /alpha exports to main entry point
