+++ workspaces/tekton/.dockerignore
@@ -2,7 +2,5 @@
 .yarn/cache
 .yarn/install-state.gz
 node_modules
-packages/*/src
-packages/*/node_modules
 plugins
 *.local.yaml
