--- workspaces/tekton/packages/app/.eslintrc.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line @backstage/no-relative-monorepo-imports -- workspace ESLint shared config
-module.exports = require('../../eslint.frontend-shared.cjs')(__dirname);
