--- workspaces/tekton/packages/app/knip-report.md
@@ -1,2 +1,0 @@
-# Knip report
-
