--- workspaces/tekton/packages/backend/knip-report.md
@@ -1,2 +1,0 @@
-# Knip report
-
