+++ workspaces/tekton/plugins/tekton/dev/index.tsx
@@ -13,223 +13,132 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { Entity } from '@backstage/catalog-model';
-import { Page, Header, TabbedLayout } from '@backstage/core-components';
-import { createDevApp } from '@backstage/dev-utils';
-import { EntityProvider } from '@backstage/plugin-catalog-react';
+import ReactDOM from 'react-dom/client';
+
+// eslint-disable-next-line @backstage/no-ui-css-imports-in-non-frontend
+import '@backstage/ui/css/styles.css';
+
+import { createApp } from '@backstage/frontend-defaults';
 import {
-  KubernetesApi,
+  ApiBlueprint,
+  createFrontendModule,
+  createFrontendPlugin,
+} from '@backstage/frontend-plugin-api';
+import { catalogApiRef } from '@backstage/plugin-catalog-react';
+import { catalogApiMock } from '@backstage/plugin-catalog-react/testUtils';
+import catalogPlugin from '@backstage/plugin-catalog/alpha';
+import {
   kubernetesApiRef,
-  kubernetesProxyApiRef,
   kubernetesAuthProvidersApiRef,
-  KubernetesProxyApi,
+  kubernetesProxyApiRef,
 } from '@backstage/plugin-kubernetes-react';
+import { AuthorizeResult } from '@backstage/plugin-permission-common';
 import { permissionApiRef } from '@backstage/plugin-permission-react';
-import { mockApis, TestApiProvider } from '@backstage/test-utils';
-import { tektonTranslations } from '@backstage-community/plugin-tekton';
+import userSettingsPlugin from '@backstage/plugin-user-settings/alpha';
 
-import { mockKubernetesPlrResponse } from '../src/__fixtures__/1-pipelinesData';
+import tektonPlugin from '../src';
+import tektonTranslationsModule from '../src/translations';
 import {
-  acsDeploymentCheck,
-  acsImageCheckResults,
-  acsImageScanResult,
-} from '../src/__fixtures__/advancedClusterSecurityData';
-import { enterpriseContractResult } from '../src/__fixtures__/enterpriseContractData';
-import { TektonCI, tektonPlugin } from '../src/plugin';
+  mockEntity,
+  mockKubernetesAuthProviderApi,
+  mockKubernetesClient,
+  mockKubernetesProxyApi,
+} from './shared';
 
-const mockEntity: Entity = {
-  apiVersion: 'backstage.io/v1alpha1',
-  kind: 'Component',
-  metadata: {
-    name: 'backstage',
-    description: 'backstage.io',
-    annotations: {
-      'backstage.io/kubernetes-id': 'backstage',
-      'tekton.dev/cicd': 'true',
-    },
-  },
-  spec: {
-    lifecycle: 'production',
-    type: 'service',
-    owner: 'user:guest',
-  },
-};
+const kubernetesStubPlugin = createFrontendPlugin({
+  pluginId: 'kubernetes',
+  extensions: [],
+});
 
-const mockPermissionApi = mockApis.permission({});
-class MockKubernetesProxyApi implements KubernetesProxyApi {
-  async getPodLogs(_request: any): Promise<any> {
-    const delayedResponse = (data: string, ms: number) =>
-      new Promise(resolve => {
-        setTimeout(() => {
-          resolve({
-            text: data,
-          });
-        }, ms);
-      });
+const kubernetesDevModule = createFrontendModule({
+  pluginId: 'kubernetes',
+  extensions: [
+    ApiBlueprint.make({
+      name: 'kubernetes-mock',
+      params: defineParams =>
+        defineParams({
+          api: kubernetesApiRef,
+          deps: {},
+          factory: () => mockKubernetesClient,
+        }),
+    }),
+    ApiBlueprint.make({
+      name: 'kubernetes-proxy-mock',
+      params: defineParams =>
+        defineParams({
+          api: kubernetesProxyApiRef,
+          deps: {},
+          factory: () => mockKubernetesProxyApi,
+        }),
+    }),
+  ],
+});
 
-    if (_request.podName.includes('ec-task')) {
-      return delayedResponse(JSON.stringify(enterpriseContractResult), 100);
-    }
+const kubernetesAuthStubPlugin = createFrontendPlugin({
+  pluginId: 'kubernetes-auth-providers',
+  extensions: [],
+});
 
-    if (_request.podName.includes('image-scan-task')) {
-      return delayedResponse(JSON.stringify(acsImageScanResult), 200);
-    }
+const kubernetesAuthDevModule = createFrontendModule({
+  pluginId: 'kubernetes-auth-providers',
+  extensions: [
+    ApiBlueprint.make({
+      name: 'kubernetes-auth-mock',
+      params: defineParams =>
+        defineParams({
+          api: kubernetesAuthProvidersApiRef,
+          deps: {},
+          factory: () => mockKubernetesAuthProviderApi,
+        }),
+    }),
+  ],
+});
 
-    if (_request.podName.includes('image-check-task')) {
-      return delayedResponse(JSON.stringify(acsImageCheckResults), 300);
-    }
+const permissionDevModule = createFrontendModule({
+  pluginId: 'app',
+  extensions: [
+    ApiBlueprint.make({
+      name: 'permission',
+      params: defineParams =>
+        defineParams({
+          api: permissionApiRef,
+          deps: {},
+          factory: () => ({
+            authorize: async () => ({ result: AuthorizeResult.ALLOW }),
+          }),
+        }),
+    }),
+  ],
+});
 
-    if (_request.podName.includes('deployment-check-task')) {
-      return delayedResponse(JSON.stringify(acsDeploymentCheck), 400);
-    }
+const catalogDevModule = createFrontendModule({
+  pluginId: 'catalog',
+  extensions: [
+    ApiBlueprint.make({
+      name: 'catalog-mock',
+      params: defineParams =>
+        defineParams({
+          api: catalogApiRef,
+          deps: {},
+          factory: () => catalogApiMock({ entities: [mockEntity] }),
+        }),
+    }),
+  ],
+});
 
-    const response = `\nstreaming logs from container: ${_request.containerName} \n...`;
-    return delayedResponse(response, 500);
-  }
+const app = createApp({
+  features: [
+    catalogPlugin,
+    userSettingsPlugin,
+    tektonPlugin,
+    tektonTranslationsModule,
+    catalogDevModule,
+    kubernetesStubPlugin,
+    kubernetesDevModule,
+    kubernetesAuthStubPlugin,
+    kubernetesAuthDevModule,
+    permissionDevModule,
+  ],
+});
 
-  async deletePod(): Promise<any> {
-    return {};
-  }
-
-  async getEventsByInvolvedObjectName(): Promise<any> {
-    return {};
-  }
-}
-class MockKubernetesClient implements KubernetesApi {
-  readonly resources;
-
-  constructor(fixtureData: { [resourceType: string]: any[] }) {
-    this.resources = Object.entries(fixtureData).flatMap(
-      ([type, resources]) => {
-        if (type === 'pipelineruns' && resources[0]?.kind === 'PipelineRun') {
-          return {
-            type: 'customresources',
-            resources,
-          };
-        } else if (type === 'taskruns' && resources[0]?.kind === 'TaskRun') {
-          return {
-            type: 'customresources',
-            resources,
-          };
-        }
-        return {
-          type: type.toLocaleLowerCase('en-US'),
-          resources,
-        };
-      },
-    );
-  }
-
-  async getWorkloadsByEntity(_request: any): Promise<any> {
-    return {
-      items: [
-        {
-          cluster: { name: 'mock-cluster' },
-          resources: this.resources,
-          podMetrics: [],
-          errors: [],
-        },
-      ],
-    };
-  }
-  async getCustomObjectsByEntity(_request: any): Promise<any> {
-    return {
-      items: [
-        {
-          cluster: { name: 'mock-cluster' },
-          resources: this.resources,
-          podMetrics: [],
-          errors: [],
-        },
-      ],
-    };
-  }
-
-  async getObjectsByEntity(): Promise<any> {
-    return {
-      items: [
-        {
-          cluster: { name: 'mock-cluster' },
-          resources: this.resources,
-          podMetrics: [],
-          errors: [],
-        },
-      ],
-    };
-  }
-
-  async getClusters(): Promise<{ name: string; authProvider: string }[]> {
-    return [{ name: 'mock-cluster', authProvider: 'serviceAccount' }];
-  }
-
-  async getCluster(_clusterName: string): Promise<
-    | {
-        name: string;
-        authProvider: string;
-        oidcTokenProvider?: string;
-        dashboardUrl?: string;
-      }
-    | undefined
-  > {
-    return { name: 'mock-cluster', authProvider: 'serviceAccount' };
-  }
-
-  async proxy(_options: { clusterName: String; path: String }): Promise<any> {
-    return {
-      kind: 'Namespace',
-      apiVersion: 'v1',
-      metadata: {
-        name: 'mock-ns',
-      },
-    };
-  }
-}
-
-const mockKubernetesAuthProviderApiRef = {
-  decorateRequestBodyForAuth: async () => {
-    return {
-      entity: {
-        apiVersion: 'v1',
-        kind: 'xyz',
-        metadata: { name: 'hey' },
-      },
-    };
-  },
-  getCredentials: async () => {
-    return {};
-  },
-};
-
-createDevApp()
-  .addTranslationResource(tektonTranslations)
-  .setAvailableLanguages(['en', 'de', 'es', 'fr', 'it', 'ja'])
-  .addPage({
-    element: (
-      <TestApiProvider
-        apis={[
-          [
-            kubernetesApiRef,
-            new MockKubernetesClient(mockKubernetesPlrResponse),
-          ],
-          [kubernetesProxyApiRef, new MockKubernetesProxyApi()],
-          [permissionApiRef, mockPermissionApi],
-          [kubernetesAuthProvidersApiRef, mockKubernetesAuthProviderApiRef],
-        ]}
-      >
-        <EntityProvider entity={mockEntity}>
-          <Page themeId="service">
-            <Header type="component — service" title="demo-sevice" />
-            <TabbedLayout>
-              <TabbedLayout.Route path="/" title="CI/CD">
-                <TektonCI />
-              </TabbedLayout.Route>
-            </TabbedLayout>
-          </Page>
-        </EntityProvider>
-      </TestApiProvider>
-    ),
-    title: 'Tekton CI',
-    path: '/tekton',
-  })
-  .registerPlugin(tektonPlugin)
-  .render();
+ReactDOM.createRoot(document.getElementById('root')!).render(app.createRoot());
