+++ workspaces/tekton/plugins/tekton/report-legacy.api.md
@@ -1,0 +1,18 @@
+## API Report File for "@backstage-community/plugin-tekton"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+import { BackstagePlugin } from '@backstage/core-plugin-api';
+import { Entity } from '@backstage/catalog-model';
+import { JSX as JSX_2 } from 'react/jsx-runtime';
+
+// @public
+export const isTektonCIAvailable: (entity: Entity) => boolean;
+
+// @public
+export const TektonCI: () => JSX_2.Element | null;
+
+// @public
+export const tektonPlugin: BackstagePlugin<{}, {}, {}>;
+```
