+++ workspaces/tekton/plugins/tekton/src/legacy.test.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 The Backstage Authors
+ * Copyright 2026 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,4 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import '@testing-library/jest-dom';
+import { tektonPlugin } from './legacy';
+
+describe('tekton legacy', () => {
+  it('should export plugin', () => {
+    expect(tektonPlugin).toBeDefined();
+  });
+});
