+++ workspaces/tekton/plugins/tekton/src/plugin.test.ts
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { tektonPlugin } from './plugin';
+import tektonPlugin from './plugin';
 
 describe('tekton', () => {
   it('should export plugin', () => {
