+++ workspaces/tekton/tsconfig.json
@@ -1,8 +1,6 @@
 {
   "extends": "@backstage/cli/config/tsconfig.json",
   "include": [
-    "packages/*/src",
-    "packages/*/config.d.ts",
     "plugins/*/src",
     "plugins/*/config.d.ts",
     "plugins/*/dev",
